refactor: extract reconfigure flow mixin from config flow
